#21 - separated BoxReader and BoxTypes
