+++ src/main/java/org/opengis/cite/gmljpx20/util/jp2/JP2Stream.java
@@ -1,10 +1,12 @@
 package org.opengis.cite.gmljpx20.util.jp2;
 
-import org.opengis.cite.gmljpx20.box.Box;
+import static org.opengis.cite.gmljpx20.util.jp2.BoxReader.readFromStream;
 
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
+
+import org.opengis.cite.gmljpx20.box.Box;
 
 public class JP2Stream {
 
@@ -16,7 +18,7 @@ public class JP2Stream {
         fs = source;
         while ( true ) {
             try {
-                Box box = Box.fromStream( source );
+                Box box = readFromStream( source );
                 boxes.add( box );
             } catch ( Exception e ) {
                 return;
